[PATCH v2] arm64: Fix page table walk of 1GB section
makedumpfile was generating large (> 500MB) vmcore files for an arm64
board with 2GB of DRAM.  It was not excluding any pages because the
mem_map address was not being converted correctly.

readmem: Can't convert a virtual address(ffffffc07fff6000) to physical
address.
readmem: type_addr: 0, addr:ffffffc07fff6000, size:16
section_mem_map_addr: Can't get a struct mem_section(ffffffc07fff6000).
mem_map (0)
mem_map : 0

makedumpfile was not handling 1GB sections in the PGD and was trying to
drill down to a PTE in which it was trying to dereference invalid
memory.  This patch adds code to check the PGD for a section type and
handle it instead of treating it as a table entry.

Signed-off-by: Bradley Bolen <bradleybolen@gmail.com>